DATAGEODE-273 - Fix DiskStoreConfiguration, EnableDiskStores annotation maxOpLogsize attribute extraction, treating the long value as an int thereby causing a ClassCastException.

+++ spring-data-geode/src/main/java/org/springframework/data/gemfire/config/annotation/DiskStoresConfiguration.java
@@ -64,7 +64,7 @@ public class DiskStoresConfiguration extends DiskStoreConfiguration {
 			enableDiskStoresAttributes.<Integer>getNumber("compactionThreshold"), 50);
 
 		setAttributeIfNotDefault(diskStoreAttributes, "maxOplogSize",
-			enableDiskStoresAttributes.<Integer>getNumber("maxOplogSize"), 1024);
+			enableDiskStoresAttributes.<Long>getNumber("maxOplogSize"), 1024L);
 
 		return diskStoreAttributes;
 	}
